refactor: 홈 페이지를 루트 경로로 이동
